feat: add Container configuration types for GitHub workflow jobs
