+++ crates/gh-workflow/src/job.rs
@@ -9,8 +9,9 @@ use serde_json::Value;
 use crate::concurrency::Concurrency;
 use crate::step::{Step, StepType, StepValue};
 use crate::{Permissions};
+use crate::{Container};
 use crate::workflow::{
-    Artifacts, Container, Defaults, Env, Expression, RetryStrategy,
+    Artifacts, Defaults, Env, Expression, RetryStrategy,
     Secret, Strategy,
 };
